requirements.txt
asgiref==3.6.0
Django==4.1.5
django-appconf==1.0.5
django-imagekit==4.1.0
djangorestframework==3.14.0
pilkit==2.0
Pillow==9.4.0
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
